requirements.txt
flask
flask-cors
pymongo
opencv-contrib-python
numpy
pillow
python-dotenv
werkzeug
requests
authlib
